# Backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Unora mobile application",
        "title": "Unora Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/Unora-app/Backend",
            "email": "support@unora.app"
        },
        "license": {
            "name": "Private",
            "url": "https://github.com/Unora-app/Backend"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/connections/{connectionId}/nudge": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send a nudge to streak partner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streaks"
                ],
                "summary": "Send nudge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Connection ID",
                        "name": "connectionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nudge message",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_streak_dto.SendNudgeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Nudge sent",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_streak_dto.NudgeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Already nudged today",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/connections/{connectionId}/streak": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get streak details for a connection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streaks"
                ],
                "summary": "Get streak",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Connection ID",
                        "name": "connectionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Streak details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_streak_dto.StreakResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Streak not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/connections/{connectionId}/streak/check-in": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Perform daily check-in for a streak",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streaks"
                ],
                "summary": "Daily check-in",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Connection ID",
                        "name": "connectionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Check-in data",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_streak_dto.CheckInRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Check-in created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_streak_dto.CheckInResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Already checked in or streak not active",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/nudges/received": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get nudges received by the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streaks"
                ],
                "summary": "Get received nudges",
                "responses": {
                    "200": {
                        "description": "Received nudges",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_UnoraApp_be_internal_streak_dto.NudgeResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/streaks/today": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all streaks requiring check-in today",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streaks"
                ],
                "summary": "Get today's streaks",
                "responses": {
                    "200": {
                        "description": "Today's streaks",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_streak_dto.TodayStreaksResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/streaks/{streakId}/recover": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Recover a broken streak using credits",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streaks"
                ],
                "summary": "Recover streak",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Streak ID",
                        "name": "streakId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Recovery request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_streak_dto.RecoverStreakRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recovery result",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_streak_dto.RecoverStreakResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Cannot recover or insufficient credits",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/streaks/{streakId}/recovery-options": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get available options for recovering a broken streak",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "streaks"
                ],
                "summary": "Get recovery options",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Streak ID",
                        "name": "streakId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recovery options",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_UnoraApp_be_internal_streak_dto.RecoveryOptionsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Streak not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_UnoraApp_be_internal_streak_dto.CheckInRequest": {
            "description": "Daily check-in request",
            "type": "object",
            "properties": {
                "activity": {
                    "type": "string",
                    "maxLength": 200,
                    "example": "Had a great video call"
                }
            }
        },
        "github_com_UnoraApp_be_internal_streak_dto.CheckInResponse": {
            "description": "Check-in record",
            "type": "object",
            "properties": {
                "activity": {
                    "type": "string",
                    "example": "Had a video call"
                },
                "checkInType": {
                    "type": "string",
                    "example": "manual"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2024-01-05T00:00:00Z"
                },
                "dayNumber": {
                    "type": "integer",
                    "example": 5
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "userId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "github_com_UnoraApp_be_internal_streak_dto.NudgeResponse": {
            "description": "Nudge between streak partners",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2024-01-05T10:00:00Z"
                },
                "dayNumber": {
                    "type": "integer",
                    "example": 5
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "message": {
                    "type": "string",
                    "example": "Don't forget to check in!"
                },
                "seenAt": {
                    "type": "string"
                },
                "senderId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "senderName": {
                    "type": "string",
                    "example": "John"
                },
                "status": {
                    "type": "string",
                    "example": "sent"
                },
                "streakId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "github_com_UnoraApp_be_internal_streak_dto.RecoverStreakRequest": {
            "description": "Request to recover a broken streak",
            "type": "object",
            "properties": {
                "payWithCredits": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_UnoraApp_be_internal_streak_dto.RecoverStreakResponse": {
            "description": "Streak recovery result",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Streak recovered successfully!"
                },
                "newCredits": {
                    "type": "integer",
                    "example": 50
                },
                "streakRestored": {
                    "type": "boolean",
                    "example": true
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "github_com_UnoraApp_be_internal_streak_dto.RecoveryOptionsResponse": {
            "description": "Streak recovery pricing options",
            "type": "object",
            "properties": {
                "canRecover": {
                    "type": "boolean",
                    "example": true
                },
                "creditCost": {
                    "type": "integer",
                    "example": 50
                },
                "recoveryDeadline": {
                    "type": "string",
                    "example": "2024-01-06T00:00:00Z"
                },
                "userCredits": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "github_com_UnoraApp_be_internal_streak_dto.SendNudgeRequest": {
            "description": "Send nudge to streak partner",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "maxLength": 200,
                    "example": "Hey! Don't forget to check in today 💪"
                }
            }
        },
        "github_com_UnoraApp_be_internal_streak_dto.StreakResponse": {
            "description": "Streak details for a connection",
            "type": "object",
            "properties": {
                "checkIns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_UnoraApp_be_internal_streak_dto.CheckInResponse"
                    }
                },
                "completedAt": {
                    "type": "string"
                },
                "connectionId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "currentDay": {
                    "type": "integer",
                    "example": 5
                },
                "healthScore": {
                    "type": "number",
                    "example": 0.85
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "myCheckInToday": {
                    "type": "boolean",
                    "example": true
                },
                "partnerCheckInToday": {
                    "type": "boolean",
                    "example": false
                },
                "recoveryDeadline": {
                    "type": "string"
                },
                "resetCount": {
                    "type": "integer",
                    "example": 0
                },
                "state": {
                    "type": "string",
                    "example": "active"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2024-01-05T00:00:00Z"
                }
            }
        },
        "github_com_UnoraApp_be_internal_streak_dto.TodayStreakItem": {
            "description": "Streak item for today's check-in list",
            "type": "object",
            "properties": {
                "connectionId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "currentDay": {
                    "type": "integer",
                    "example": 5
                },
                "needsCheckIn": {
                    "type": "boolean",
                    "example": true
                },
                "partnerCheckedIn": {
                    "type": "boolean",
                    "example": false
                },
                "partnerName": {
                    "type": "string",
                    "example": "John"
                },
                "partnerPhotoUrl": {
                    "type": "string",
                    "example": "https://storage.example.com/photo.jpg"
                },
                "state": {
                    "type": "string",
                    "example": "active"
                },
                "streakId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "github_com_UnoraApp_be_internal_streak_dto.TodayStreaksResponse": {
            "description": "Today's streaks requiring check-in",
            "type": "object",
            "properties": {
                "streaks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_UnoraApp_be_internal_streak_dto.TodayStreakItem"
                    }
                },
                "totalCompleted": {
                    "type": "integer",
                    "example": 1
                },
                "totalPending": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "github_com_UnoraApp_be_pkg_response.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_UnoraApp_be_pkg_response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_UnoraApp_be_pkg_response.APIError"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}